updated struct
